requirements.txt
httpx==0.27.0
python-dotenv==1.0.0
pandas==2.2.0
openpyxl==3.1.2
duckduckgo-search==4.4.3
rich==13.7.0
yfinance==0.2.36
sentence-transformers==2.3.1
numpy==1.26.3
scikit-learn==1.4.0
gradio==4.16.0
redis==5.0.1